Don't double-iterate during MultiDict init (#3)
Previously, the MultiDict constructor iterated twice over its argument:
once as a sanity check on its contents, and once again to add those
contents to the dictionary. If a generator was passed as an argument,
the first iteration would drain it, and the resulting dictionary would
be empty.

To fix this, save the contents of the generator in another list. This
fixes #2.