Implement a custom parser for METH_FASTCALL | METH_KEYWORDS protocol (#1070)
The story is:

Back in the day, multidict used the `METH_VARARGS | METH_KEYWORDS`
calling protocol for its methods.
Later, the library switched to `METH_FASTCALL | METH_KEYWORDS` protocol.
The new method is significantly faster than previously used, especially
if arguments are passed as positional.

Unfortunately, parsing FASTCALL arguments is not a trivial task; and
Python doesn't provide a public parser API.
Thus, `multidict` used to utilize a private
`_PyArg_ParseStackAndKeywords` / `_PyArg_UnpackKeywords`.
The downside is obvious: a private API doesn't guarantee to work in
future Python versions.
After a segfault with Python 3.12
(https://github.com/aio-libs/multidict/issues/926),
https://github.com/aio-libs/multidict/pull/929 patch was applied to
switch back to `METH_VARARGS | METH_KEYWORDS` as a temporary fix.

While making a generic FASTCALL parser is challenging, multidict doesn't
need it, honestly. The library's methods are called with two arguments
(`.add()`), the second arg often has a default value (`.get()`). A
parser for such a case is simple; the PR provides it. The parser doesn't
depend on any non-public function, and it is compatible with any future
Python versions as FASTCALL is a fixed-in-stone calling protocol.

---------

Co-authored-by: J. Nick Koston <nick@koston.org>