Bring runtime code coverage metrics to 100%
PR #1077

Among other things, this patch excludes the pickle generator from measurements and adds a temporary "no cover" pragma to a line that's currently unreachable. That will have to be reverted as a part of the future #837-related refactoring. The type preciseness expectation has been lowered to 85% due to the ecosystem limitations.

Co-authored-by: 🇺🇦 Sviatoslav Sydorenko (Святослав Сидоренко) <webknjaz@redhat.com>
Co-authored-by: Sam Bull <git@sambull.org>