Stable benchmarks, fifth attempt (#1209)
See #1201 for details

The PR increases the keys space.
The tested idea is: if instability is because of colliding keys,
increasing the number of keys reduces the chance of getting a very bad
or very good test setup. The test setup becomes 'neutral' with more or
less predictable execution time.